cleanup the error dialog system
